Begynt smått å organisere koden basert på funksjonalitet
slettet litt kode
renamet litt kode
flyttet litt kode
