+++ src/main/kotlin/no/nav/eessi/pensjon/services/person/PersonV3Config.kt
@@ -1,4 +1,4 @@
-package no.nav.eessi.pensjon.services.personv3
+package no.nav.eessi.pensjon.services.person
 
 import no.nav.tjeneste.virksomhet.person.v3.binding.PersonV3
 import org.apache.cxf.jaxws.JaxWsProxyFactoryBean
